Moving model converters code from decay to modeling submodule
